src/agnoprj/requirements.txt
requests
agno
fastapi
uvicorn
openai
pydantic
python-dotenv
sqlalchemy
aiosqlite